- Add explicit method to read output params from a BatchResult. Can only be called after close().
- Add a method to QueryResultStream to read output params. Can only be called at the end of the stream.
- Make the row count from DML a u64 consistently instead of i64 in some places.

----
#### AI description  (iteration 1)
#### PR Classification
This PR is an API change that separates the retrieval of return values from closing a BatchResult to support reading both row data and return values from query streams.

#### PR Summary
The changes refactor how tokens are processed in query results and adjust the API to allow output parameters to be retrieved after stream consumption.
- **`tds-x/src/query/result.rs`**: Introduces a loop in token reading that handles `ReturnValue` tokens separately, adds a new method `retrieve_output_params`, and modifies `close()` to no longer return output parameters directly.
- **`tds-x/tests/test_rpc_results.rs`**: Updates test cases to use the new API for output parameter retrieval and validates the stored procedure behavior.
- **`tds-x/src/connection/tds_connection.rs` & `cpp-driver-cxx`**: Adjusts the closing workflow and type representations (switch from `i64` to `u64` for Dml results).
<!-- GitOpsUserAgent=GitOps.Apps.Server.pullrequestcopilot -->

Related work items: #37690

+++ tds-x/tests/common/mod.rs
@@ -79,7 +79,7 @@ pub async fn assert_matches_expected(qrt: QueryResultType<'_>, expected: &Expect
             QueryResultType::DmlResult(rows_affected),
             ExpectedQueryResultType::Update(expected_row_count),
         ) => {
-            assert_eq!(rows_affected, *expected_row_count as i64);
+            assert_eq!(rows_affected, *expected_row_count);
         }
     }
 }
